ref(browser): Restructure browser-utils by domain
Reorganize browser-utils/src away from the catch-all metrics/ folder into
clear domains:

- instrumentation/ for event-source hooks (dom, history, location, xhr), plus
  the PerformanceObserver layer renamed from metrics/instrument.ts to
  instrumentation/performanceObserver.ts (removing the collision with the
  instrument/ folder)
- web-vitals/ for everything web-vital: tracking, spans, inp, lcp,
  softNavCorrelation, reportEvents, and the browser helpers merged into a
  single utils.ts
- performance/ for generic performance-entry enrichment: entries, element/
  user/resource timing, and shared utils

The 907-line browserMetrics.ts is split into web-vitals/tracking.ts (web-vital
tracking) and performance/entries.ts (perf-entry/long-task/navigation/resource
spans), and utils.ts into web-vitals/reportEvents.ts and performance/utils.ts.

Public API surface (index.ts re-exports) is unchanged; only internal paths and
the test tree move.
