+++ packages/browser-utils/src/instrumentation/performanceObserver.ts
