+++ packages/browser-utils/src/instrumentation/xhr.ts
