+++ packages/browser-utils/src/performance/resourceTiming.ts
