+++ packages/browser-utils/src/web-vitals/lcp.ts
