+++ packages/browser-utils/test/instrumentation/dom.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { instrumentDOM } from '../../src/instrument/dom';
+import { instrumentDOM } from '../../src/instrumentation/dom';
 import { WINDOW } from '../../src/types';
 
 // @ts-expect-error - idk
